+++ General/FrameServer.vb
@@ -61,39 +61,9 @@ Public Structure ServerInfo
     Public FrameRateNum As Integer
     Public FrameRateDen As Integer
     Public FrameCount As Integer
+    Public ColorSpace As ColorSpace
 
-
-    ' avs2pipemod64.exe -info   
-    ' avisynth_version 2.600 / AviSynth+ 3.4 (r2923, 3.4, x86_64)
-    ' script_name      D:\Samples\Jill_temp\Jill_new.avs
-    ' 
-    ' v:width            1920
-    ' v:height           1080
-    ' v:image_type       framebased
-    ' v:field_order      assumed bottom field first
-    ' v:pixel_type       YV12
-    ' v:bit_depth        8
-    ' v:number of planes 3
-    ' v:fps              30000/1001
-    ' v:frames           6172
-    ' v:duration[sec]    205.939
-
-
-    ' vspipe.exe --info file.vpy -
-    ' Width: 1920
-    ' Height: 1080
-    ' Frames: 6172
-    ' FPS: 30000/1001 (29.970 fps)
-    ' Format Name: YUV420P8
-    ' Color Family: YUV
-    ' Alpha: No
-    ' Sample Type: Integer
-    ' Bits: 8
-    ' SubSampling W: 1
-    ' SubSampling H: 1
-
-    'TODO: add colorspace support
-    Function GetText(position As Integer) As String
+    Function GetInfoText(position As Integer) As String
         Dim rate = FrameRateNum / FrameRateDen
         Dim currentDate = Date.Today.AddSeconds(position / rate)
         Dim lengthtDate = Date.Today.AddSeconds(FrameCount / rate)
@@ -110,6 +80,45 @@ Public Structure ServerInfo
                "Height    : " & Height & BR &
                "Frames    : " + frames + BR +
                "Time      : " + len + BR +
-               "Framerate : " + rate.ToString.Shorten(9) & " (" & FrameRateNum & "/" & FrameRateDen & ")"
+               "Framerate : " + rate.ToString.Shorten(9) + " (" & FrameRateNum & "/" & FrameRateDen & ")" + BR +
+               "Format    : " + ColorSpace.ToString.Replace("_", "")
     End Function
 End Structure
+
+Public Enum ColorSpace
+    Unknown = 0
+    BGR32 = 1342177282
+    RGBP8 = -1879048191
+    RGBP10 = -1878720511
+    RGBP12 = -1878654975
+    RGBP14 = -1878589439
+    RGBP16 = -1878982655
+    Y8 = -536870912
+    Y10 = -536543232
+    Y12 = -536477696
+    Y14 = -536412160
+    Y16 = -536805376
+    Y32 = -536739840
+    YUV410P8 = -1610612471
+    YUV411P8 = -1610611959
+    YUV420P8 = -1610612720
+    YUV420P8_ = -1610612728
+    YUV420P10 = -1610285048
+    YUV420P12 = -1610219512
+    YUV420P14 = -1610153976
+    YUV420P16 = -1610547192
+    YUV420PS = -1610481656
+    YUV422P8 = -1610611960
+    YUV422P10 = -1610284280
+    YUV422P12 = -1610218744
+    YUV422P14 = -1610153208
+    YUV422P16 = -1610546424
+    YUV422PS = -1610480888
+    YUV444P8 = -1610611957
+    YUV444P10 = -1610284277
+    YUV444P12 = -1610218741
+    YUV444P14 = -1610153205
+    YUV444P16 = -1610546421
+    YUV444PS = -1610480885
+    YUY2 = 1610612740
+End Enum